fix: move DecodeableMap to a new file
